Update SignInNavigator
SignInNavigator와 MapDrawerNavigator를 Navigator로 합쳤습니다.
