+++ TT/App.js
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
 
-import SignInNavigator from './src/Navigator/SignInNavigator';
+import Navigator from './src/Navigator/Navigator';
 
 const Stack = createStackNavigator();
 
@@ -10,7 +10,7 @@ function App() {
   return (
     <NavigationContainer>
       <Stack.Navigator>
-        <Stack.Screen name="SignInNav" component={SignInNavigator} />
+        <Stack.Screen name="Navigator" component={Navigator} />
       </Stack.Navigator>
     </NavigationContainer>
   );
